pyproject.toml
[project]
name = "dteam-django-practical-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikhail Gomziak",email = "mikhailgomziak@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "weasyprint (>=65.1,<66.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg (>=3.2.7,<4.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "openai (>=1.78.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
